refactor(watchos): file structure
